Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
django-filter==23.4
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
stripe==7.8.0